sketch of keyboard manager
