+++ config.ini
@@ -1,19 +1,17 @@
 [options]
 shunt_ohms = 0.1
-base_voltage = 5.0
-maximum_value = 1024.0
-zero_level_offset = 40.0
 buttons_tolerance = 20.0
-long_press_time = 800.0
+maximum_value = 10000.0
+maximum_voltage = 5.0
+minimum_voltage = 0.0
 double_click_interval = 100.0
+long_press_time = 800.0
 read_delay = 21.0
-verbose = True
-base_value = 850.0
 
 [buttons]
-vol+ = 103
-next = 40
-prev = 300
+vol+ = 2328
+next = 4032
+prev = 6648
 
 [functions]
 vol+ = 23 23 26
