+++ keyboard/utils.py
@@ -1,0 +1,3 @@
+from keyboard.keyboard_manager import KeyboardManager
+
+keyboard = KeyboardManager()
